vuex_router_sync: rename store file
